server/pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = "Server for vscode decompiler."
authors = ["novafacing <rowanbhart@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
grpcio = "^1.46.3"
protobuf = "^4.21.1"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
isort = "^5.10.1"
mypy = "^0.961"
pylint = "^2.14.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pylint.messages_control]
max-line-length = 88
disable = ["attribute-defined-outside-init", "logging-fstring-interpolation"]

[tool.pylint.basic]
argument-naming-style = "snake_case"
attr-naming-style = "snake_case"
class-attribute-naming-style = "snake_case"
class-naming-style = "PascalCase"
const-naming-style = "UPPER_CASE"
function-naming-style = "snake_case"
method-naming-style = "snake_case"
module-naming-style = "snake_case"
variable-naming-style = "any"

[tool.pylint.format]
expected-line-ending-format = "LF"
max-module-lines = 600
logging-format-style = "new"
max-args = 6
max-attributes = 12
max-bool-expr = 4
max-locals = 16
max-parents = 7
max-public-methods = 16
max-statements = 64
min-public-methods = 1
good-names = ["e"]

[tool.pylint.exceptions]
overgeneral-exceptions = "Exception"

[tool.black]
line-length = 88
target-version = ["py36", "py37", "py38", "py39"]
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | tests/.*/setup.py
)/
'''

[tool.mypy]
follow_imports = "normal"
disallow_any_unimported = true
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
strict_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
warn_unreachable = true
strict_equality = true

[tool.isort]
profile = "black"
multi_line_output = 3
use_parentheses = true